Consolidate TestEndpoint
