+++ dialogue-blocking-channels/src/test/java/com/palantir/dialogue/blocking/BlockingChannelAdapterTest.java
@@ -23,15 +23,12 @@ import static org.mockito.Mockito.verify;
 import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.Uninterruptibles;
 import com.palantir.dialogue.Channel;
-import com.palantir.dialogue.Endpoint;
-import com.palantir.dialogue.HttpMethod;
 import com.palantir.dialogue.Request;
 import com.palantir.dialogue.Response;
+import com.palantir.dialogue.TestEndpoints;
 import com.palantir.dialogue.TestResponse;
-import com.palantir.dialogue.UrlBuilder;
 import com.palantir.logsafe.exceptions.SafeRuntimeException;
 import java.time.Duration;
-import java.util.Map;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.atomic.AtomicBoolean;
@@ -41,32 +38,6 @@ import org.junit.jupiter.api.Test;
 public class BlockingChannelAdapterTest {
 
     private final Response stubResponse = new TestResponse().code(200);
-
-    private final Endpoint stubEndpoint = new Endpoint() {
-
-        @Override
-        public void renderPath(Map<String, String> _params, UrlBuilder _url) {}
-
-        @Override
-        public HttpMethod httpMethod() {
-            return HttpMethod.POST;
-        }
-
-        @Override
-        public String serviceName() {
-            return "service";
-        }
-
-        @Override
-        public String endpointName() {
-            return "endpoint";
-        }
-
-        @Override
-        public String version() {
-            return "1.0.0";
-        }
-    };
 
     @Test
     public void testSuccessful() {
@@ -80,7 +51,7 @@ public class BlockingChannelAdapterTest {
             return stubResponse;
         });
         ListenableFuture<Response> result =
-                channel.execute(stubEndpoint, Request.builder().build());
+                channel.execute(TestEndpoints.GET, Request.builder().build());
         assertThat(result).isNotDone();
         latch.countDown();
         Awaitility.waitAtMost(Duration.ofSeconds(3)).untilAsserted(() -> {
@@ -95,7 +66,7 @@ public class BlockingChannelAdapterTest {
             throw new SafeRuntimeException("expected");
         });
         ListenableFuture<Response> result =
-                channel.execute(stubEndpoint, Request.builder().build());
+                channel.execute(TestEndpoints.GET, Request.builder().build());
         Awaitility.waitAtMost(Duration.ofSeconds(3)).until(result::isDone);
         assertThatThrownBy(result::get)
                 .isInstanceOf(ExecutionException.class)
@@ -116,7 +87,7 @@ public class BlockingChannelAdapterTest {
             return response;
         });
         ListenableFuture<Response> result =
-                channel.execute(stubEndpoint, Request.builder().build());
+                channel.execute(TestEndpoints.GET, Request.builder().build());
         channelLatch.await();
         assertThat(result.cancel(true)).isTrue();
         assertThat(result).isCancelled();
