+++ dialogue-core/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     testImplementation 'org.mockito:mockito-junit-jupiter'
     testRuntimeOnly 'org.apache.logging.log4j:log4j-slf4j-impl'
     testRuntimeOnly 'org.apache.logging.log4j:log4j-core'
+    testImplementation project(':dialogue-test-common')
 
     annotationProcessor 'org.immutables:value'
     testAnnotationProcessor 'org.immutables:value'
